apiecoforesthub/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
PyJWT==2.8.0
sqlalchemy==2.0.23
mysql-connector-python== 8.2.0
mysqlclient == 2.2.0
bcrypt==4.0.1
google-cloud-bigquery==3.13.0